Split out daf yomi into separate module
