+++ hdate/htables.py
@@ -1,7 +1,5 @@
 """Constant lookup tables for hdate modules."""
 
-import datetime as dt
-from dataclasses import dataclass
 from enum import Enum, IntEnum, auto
 
 from hdate.translator import TranslatorMixin
@@ -347,60 +345,3 @@ PARASHA_SEQUENCES: dict[tuple[int, ...], tuple[Enum, ...]] = {
         *erange(Parasha.DEVARIM, Parasha.NITZAVIM),
     ),
 }
-
-
-# The first few cycles were only 2702 blatt. After that it became 2711. Even with
-# that, the math doesn't play nicely with the dates before the 11th cycle :(
-# From cycle 11 onwards, it was simple and sequential
-DAF_YOMI_CYCLE_11_START = dt.date(1997, 9, 29)
-
-
-@dataclass
-class Masechta(TranslatorMixin):
-    """Masechta object."""
-
-    name: str
-    pages: int
-
-
-DAF_YOMI_MESECHTOS = (
-    Masechta("berachos", 63),
-    Masechta("shabbos", 156),
-    Masechta("eruvin", 104),
-    Masechta("pesachim", 120),
-    Masechta("shekalim", 21),
-    Masechta("yoma", 87),
-    Masechta("succah", 55),
-    Masechta("beitzah", 39),
-    Masechta("rosh_hashanah", 34),
-    Masechta("taanis", 30),
-    Masechta("megillah", 31),
-    Masechta("moed_katan", 28),
-    Masechta("chagigah", 26),
-    Masechta("yevamos", 121),
-    Masechta("kesubos", 111),
-    Masechta("nedarim", 90),
-    Masechta("nazir", 65),
-    Masechta("sotah", 48),
-    Masechta("gittin", 89),
-    Masechta("kiddushin", 81),
-    Masechta("bava_kamma", 118),
-    Masechta("bava_metzia", 118),
-    Masechta("bava_basra", 175),
-    Masechta("sanhedrin", 112),
-    Masechta("makkos", 23),
-    Masechta("shevuos", 48),
-    Masechta("avodah_zarah", 75),
-    Masechta("horayos", 13),
-    Masechta("zevachim", 119),
-    Masechta("menachos", 109),
-    Masechta("chullin", 141),
-    Masechta("bechoros", 60),
-    Masechta("arachin", 33),
-    Masechta("temurah", 33),
-    Masechta("kereisos", 27),
-    Masechta("meilah", 36),
-    Masechta("niddah", 72),
-)
-
-DAF_YOMI_TOTAL_PAGES = sum(mesechta.pages for mesechta in DAF_YOMI_MESECHTOS)
